预约相关页面完善，我的预约页面调整
预约相关页面完善，我的预约页面调整

+++ ME时代/Classes/Main/Store/Controller/MEServiceDetailsVC.m
@@ -151,7 +151,7 @@ kTDWebViewCellDidFinishLoadNotificationMethod
 }
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath{
-    NSLog(@"%ld",indexPath.section);
+    NSLog(@"%ld",(long)indexPath.section);
     
 }
